Move traits to providers/openAI
